refactor: APIクライアントのディレクトリ変更に伴うスキーマの修正
